move common/config_file into classes folder
