+++ slk/chain/xchain_transfer.py
@@ -3,7 +3,7 @@ import time
 from xrpl.models import Amount, Memo, Payment
 
 from slk.chain.chain import Chain
-from slk.common import Account
+from slk.classes.common import Account
 from slk.sidechain_params import SidechainParams
