+++ slk/classes/common.py
