+++ tests/utils.py
@@ -9,7 +9,7 @@ from tabulate import tabulate
 from xrpl.models import XRP, Amount, IssuedCurrencyAmount, Subscribe
 
 from slk.chain.chain import Chain, balances_data
-from slk.common import Account, same_amount_new_value
+from slk.classes.common import Account, same_amount_new_value
 
 MC_SUBSCRIBE_QUEUE = []
 SC_SUBSCRIBE_QUEUE = []
